requirements.txt
# python==3.11.5
pyyaml==6.0.1
# pandas==2.1.0
# numpy==1.25.2
# xarray==2024.1.1
# h5netcdf==1.3.0
# scipy==1.11.2
# matplotlib==3.7.2
# geopandas==0.13.2
# seaborn==0.13.2

pandas==2.0.3
numpy
xarray==2023.1.0
h5netcdf
scipy
matplotlib
# geopandas
seaborn